House keeping: split source code into different files
