+++ ring_doorbell/stickup_cam.py
@@ -1,0 +1,17 @@
+# coding: utf-8
+# vim:sw=4:ts=4:et:
+"""Python Ring Doorbell wrapper."""
+import logging
+
+from ring_doorbell import RingDoorBell
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class RingStickUpCam(RingDoorBell):
+    """Implementation for RingStickUpCam."""
+
+    @property
+    def family(self):
+        """Return Ring device family type."""
+        return 'stickup_cams'
